Start out framework for modules
Modules will include base.h and define the functions there.
A make script will be created to take the modules and combine them with
module.c, where the main function is. If it errors at any time, it possibly
uses status to print the error and retrun an appropriate code. Otherwise, it
prints a string to stdout and returns SUCCESS