Start implementation of configuration manager
conf.? -- residing in the lib folder -- which parses command line arguments,
loads configuration files, and returns commonly used variables. In addition,
it builds a VMap for the quincyrc file.

The quincyrc (default config file name) supports a format that I'm guessing is
similar to git. [directives] set the scope for the following variables, and
varibales are in a
	variable = value
format.

The built configuration map is not currently used, but it will be the second
source for variables after the command line arguments. Basically variables
defined in later stages overwrite the previous ones:
	compile time defines
	loaded from quincyrc
	specified on the command line

The test/ folder has been created, with a simple program showing off what the
settings are using various CLI parameters and quincyrc variables

lib/defines.h is the compile time constant definition
The included .def's are examples of the corresponding file.