Add an uncommented mess. See full commit log
Whenever we get a PRIVMSG, we first try to match against the command regex. If
that matches, then we set a variable saying that it matched (for the
"specific" uargs), and alter the matchable string to just the end.

We then loop through all modules, and if they match against the leftover
string we _exec them and print the output back to whatever channel we're in.
This does not handle private messages properly.